Now uploads and updates to google calendar
Uploads and Updates to google calendar.
Some code is a bit scruffy and I want to have a quick go over to make it a bit more tidy.

Other then that, happy with the project and didn't take very long.

Possibly make into a mobile app or a desktop app where you can choose sport, teams and leagues. But thats in the future.